requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
executing==2.0.1
fastapi==0.108.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.2
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.0
llama_cpp_python==0.2.26
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.3
openai==1.6.1
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.32.0.post1
starlette-context==0.3.6
telepot==12.7
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.12
yarl==1.9.4
